Fix overflow in raw bytes counter and fix compiler warnings created by that fix
git-svn-id: svn+ssh://svn.code.sf.net/p/turbovnc/code/trunk@1856 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b

+++ unix/Xvnc/programs/Xserver/hw/vnc/rfb.h
@@ -256,16 +256,16 @@ typedef struct rfbClientRec {
 
     /* statistics */
 
-    int rfbBytesSent[MAX_ENCODINGS];
+    long long rfbBytesSent[MAX_ENCODINGS];
     int rfbRectanglesSent[MAX_ENCODINGS];
     int rfbLastRectMarkersSent;
-    int rfbLastRectBytesSent;
-    int rfbCursorShapeBytesSent;
+    long long rfbLastRectBytesSent;
+    long long rfbCursorShapeBytesSent;
     int rfbCursorShapeUpdatesSent;
-    int rfbCursorPosBytesSent;
+    long long rfbCursorPosBytesSent;
     int rfbCursorPosUpdatesSent;
     int rfbFramebufferUpdateMessagesSent;
-    int rfbRawBytesEquivalent;
+    long long rfbRawBytesEquivalent;
     int rfbKeyEventsRcvd;
     int rfbPointerEventsRcvd;
